requirements.txt
langchain>=0.0.208
langchain-openai>=0.0.2
openai>=1.3.0
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.28.2
pinecone-client>=2.2.4
chromadb>=0.4.18
unstructured>=0.10.30
tiktoken>=0.5.1
mem0ai>=0.0.5
psycopg2-binary>=2.9.6
pgvector>=0.2.2
langchain-postgres>=0.0.1
sqlalchemy>=2.0.0
clickup-python
python-multipart>=0.0.6

# Document processing
PyPDF2>=3.0.0
pytesseract>=0.3.10
pdf2image>=1.16.3
Pillow>=9.5.0
python-docx>=0.8.11
openpyxl>=3.1.2
pandas>=2.0.3
docling>=0.1.0

# NLP for entity extraction
spacy>=3.6.1

# Web app
Jinja2>=3.1.2 